tests/wpt/web-platform-tests/dom/nodes/Document-getElementsByClassName.html

<!DOCTYPE html>
<title>Document.getElementsByClassName</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var a = document.createElement("a"),
      b = document.createElement("b");
  a.className = "foo";
  this.add_cleanup(function() {document.body.removeChild(a);});
  document.body.appendChild(a);

  var l = document.getElementsByClassName("foo");
  assert_true(l instanceof HTMLCollection);
  assert_equals(l.length, 1);

  b.className = "foo";
  document.body.appendChild(b);
  assert_equals(l.length, 2);

  document.body.removeChild(b);
  assert_equals(l.length, 1);
}, "getElementsByClassName() should be a live collection");
</script>
